提交或者回滚时直接释放连接

+++ lorne-tx-core/src/main/java/com/lorne/tx/db/AbstractConnection.java
@@ -94,8 +94,15 @@ public abstract class AbstractConnection implements Connection {
         }
         logger.info("close-state->" + state + "," + groupId);
         if (state == 0) {
-            closeConnection();
-            dataSourceService.deleteCompensateId(transactionLocal.getCompensateId());
+            //再嵌套时，第一次成功后面出现回滚。
+            if(waitTask.isAwait()&&!waitTask.isRemove()) {
+                //通知第一个连接回滚事务。
+                waitTask.setState(0);
+                waitTask.signalTask();
+            }else {
+                closeConnection();
+                dataSourceService.deleteCompensateId(transactionLocal.getCompensateId());
+            }
         }
         if (state == 1) {
